+++ lib/mailgun_logger_web/views/page_view.ex
@@ -1,4 +1,10 @@
 defmodule MailgunLoggerWeb.PageView do
   use MailgunLoggerWeb, :view
   @moduledoc false
+
+  def print_date_labels(days) do
+    days
+    |> Enum.map(& Calendar.strftime(&1, "%Y-%m-%d %H:%M")) 
+    |> Jason.encode!()
+  end
 end